fix include file
    コンパイル時にエラーが出ていたが、原因を特定し、修正した。
    原因は、パラメータサーバーからの値の読み出しはすべてdoubleでなければなら
    ないが、それをfloatにしてしまっていたために起きていた。
    よってこれを修正した。

+++ include/gim30/gim30_node_control.hpp
@@ -19,8 +19,8 @@ namespace Gim30{
     
     void GetAngle();
     
-    float old_angle;
-    float new_angle;
+    double old_angle;
+    double new_angle;
 
   private:
     void SetGim30();
